Calculate the GeoCBI from input variables

Calculate the GeoCBI from input variables

calculate_geocbi(
  s1_fcov_post_fire_leaf_fall = NA_real_,
  s1_area_unburnt = NA_real_,
  s1_duff = NA_real_,
  s1_coarse_fuel = NA_real_,
  s2_area_unburnt = NA_real_,
  s2_grass_trees_with_skirts = NA_real_,
  s2_unburnt_shrub_density = NA_real_,
  s2_fcov_regenerating_plants = NA_real_,
  s3_fcov_original_crown_intact = NA_real_,
  s3_density_bare_shrubs = NA_real_,
  s4_fcov_original_crown_intact = NA_real_,
  s4_resprouting_on_live_trees = NA_real_,
  s4_char_height = NA_real_,
  s5_fcov_original_crown_intact = NA_real_,
  s5_no_trees_over_20cm_fallen = NA_real_,
  s5_resprouting_on_live_trees = NA_real_,
  s5_char_height = NA_real_,
  s1_fcov = 1,
  s2_fcov = 0,
  s3_fcov = 0,
  s4_fcov = 0,
  s5_fcov = 0,
  verbose = FALSE
)

Arguments

s1_fcov_post_fire_leaf_fall
A numeric grade from 0.0 to 3.0. Post-fire leaf fall cover. Includes leaves that have fallen after the burn but not unburned patches.

s1_area_unburnt
A numeric grade from 0.0 to 3.0. Unburnt area. The approximate percentage of area not burned.

s1_duff
A numeric grade from 0.0 to 3.0. Duff condition (crushed sticks and leaves). Broken leaf pieces that form a type of mulch under the litter layer. If necessary, scrape down to mineral soil to see how deeply the char has penetrated.

s1_coarse_fuel
A numeric grade from 0.0 to 3.0. Coarse fuel condition (logs). Logs greater than 6 mm diameter that were on the ground prior to the burn.

s2_area_unburnt
A numeric grade from 0.0 to 3.0. Unburnt area. The approximate percentage of area not burned.

s2_grass_trees_with_skirts
A numeric grade from 0.0 to 3.0. Grass trees with skirts. How many grass trees still have skirts?

s2_unburnt_shrub_density
A numeric grade from 0.0 to 3.0. Density of unburnt shrubs. When walking across plot, how would you touch shrubs below 1m?

s2_fcov_regenerating_plants
A numeric grade from 0.0 to 3.0. Regenerating plants cover. All green growth associated at the base of shrubs or apparently out in the open. Include seedlings in the approximate area.

s3_fcov_original_crown_intact
A numeric grade from 0.0 to 3.0. Intact original crown cover. How much original crown is intact?

s3_density_bare_shrubs
A numeric grade from 0.0 to 3.0. Density of bare shrubs > 1m.

s4_fcov_original_crown_intact
A numeric grade from 0.0 to 3.0. Intact original crown cover. How much original crown is intact?

s4_resprouting_on_live_trees
A numeric grade from 0.0 to 3.0. Resprouting on live trees.

s4_char_height
A numeric grade from 0.0 to 3.0. Char height. Fraction of total stratum height charred.

s5_fcov_original_crown_intact
A numeric grade from 0.0 to 3.0. Intact original crown cover. How much original crown is intact?

s5_no_trees_over_20cm_fallen
A numeric grade from 0.0 to 3.0. Number of trees > 20cm fallen.

s5_resprouting_on_live_trees
A numeric grade from 0.0 to 3.0. Resprouting on live trees.

s5_char_height
A numeric grade from 0.0 to 3.0. Char height. Fraction of total stratum height charred.

s1_fcov
The fraction of coverage of stratum 1, default: 1. This value is never captured in the digital form as it always is 1 (100%). The variable is however provided here to allow different values.

s2_fcov
The fraction of coverage of stratum 2 in quarter steps from 0.0 to 1.0. Default: 0.

s3_fcov
The fraction of coverage of stratum 2 in quarter steps from 0.0 to 1.0. Default: 0.

s4_fcov
The fraction of coverage of stratum 2 in quarter steps from 0.0 to 1.0. Default: 0.

s5_fcov
The fraction of coverage of stratum 2 in quarter steps from 0.0 to 1.0. Default: 0.

verbose
Whether to display diagnostic messages, default: FALSE.

Details

The GeoCBI is calculated from the following components:

Stratum 1 fraction of cover: always 1 (100%), therefore never captured.

Stratum 2..5 fcov: the extracted gradings of surface cover in quarter steps from 0 to 1: 0.0, 0.25, 0.5, 0.75, 1.0

Stratum scores (CBI): The average grading from all non-null gradings out of all captured variables for each stratum. The variables are named with a stratum prefix. E.g., all variables for stratum 1 are prefixed s1_.

Overall index: GeoCBI

The GeoCBI is calculated as:

(sum of stratum scores) / (sum of stratum fraction of cover)

Vars not used in GeoCBI formula

Some variables are captured, but not used for GeoCBI calculation. They provide context and metadata, and include representative photos.

"id" "observation_start_time" "reporter" "device_id" "observation_end_time" "submission_date" "submitter_id" "submitter_name" "instance_id" "location_longitude" "location_latitude" "location_altitude" "plot_photo" "plot_name" "homogeneous_burn_severity" "s1_photo" "s2_dominant_vegetation" "s2_photo" "s3_dominant_vegetation" "s3_photo" "s4_dominant_vegetation" "s4_photo" "s5_dominant_vegetation" "s5_photo" "no_grass_trees" "no_hollow_logs" "no_standing_trees_gt20cm" "no_standing_trees_with_hollows" "depth_leaf_litter_cm" "avg_gap_between_unburnt_patches" "evidence_of_native_fauna" "evidence_of_feral_fauna" "odata_context"

See also

Examples

# With missing variables
calculate_geocbi(
  s1_fcov_post_fire_leaf_fall = 1,
  s2_area_unburnt = 2.5,
  s2_fcov = 0.5,
  verbose = TRUE)

#> Stratum 1: CBI 1 * FCOV 1 = Score 1

#> Stratum 2: CBI 2.5 * FCOV 0.5 = SCore 1.25

#> Stratum 3: CBI NaN * FCOV 0 = SCore NaN

#> Stratum 4: CBI NaN * FCOV 0 = SCore NaN

#> Stratum 5: CBI NaN * FCOV 0 = Score NaN

#> GeoCBI: 1.5 = Score sum 2.25 / FCOV sums 1.5

#> [1] 1.5

# With complete variables, all set to 1
calculate_geocbi(
  s1_fcov_post_fire_leaf_fall = 1,
  s1_area_unburnt = 1,
  s1_duff = 1,
  s1_coarse_fuel = 1,
  s2_area_unburnt = 1,
  s2_grass_trees_with_skirts = 1,
  s2_unburnt_shrub_density = 1,
  s2_fcov_regenerating_plants = 1,
  s3_fcov_original_crown_intact = 1,
  s3_density_bare_shrubs = 1,
  s4_fcov_original_crown_intact = 1,
  s4_resprouting_on_live_trees = 1,
  s4_char_height = 1,
  s5_fcov_original_crown_intact = 1,
  s5_no_trees_over_20cm_fallen = 1,
  s5_resprouting_on_live_trees = 1,
  s5_char_height = 1,
  s1_fcov = 1,
  s2_fcov = 1,
  s3_fcov = 1,
  s4_fcov = 1,
  s5_fcov = 1,
  verbose = TRUE)

#> Stratum 1: CBI 1 * FCOV 1 = Score 1

#> Stratum 2: CBI 1 * FCOV 1 = SCore 1

#> Stratum 3: CBI 1 * FCOV 1 = SCore 1

#> Stratum 4: CBI 1 * FCOV 1 = SCore 1

#> Stratum 5: CBI 1 * FCOV 1 = Score 1

#> GeoCBI: 1 = Score sum 5 / FCOV sums 5

#> [1] 1